moved wind speed to atmosphere
