+++ docs/sphinx/source/reference/airmass_atmospheric.rst
@@ -17,3 +17,4 @@ Airmass and atmospheric models
    atmosphere.kasten96_lt
    atmosphere.angstrom_aod_at_lambda
    atmosphere.angstrom_alpha
+   atmosphere.windspeed_hellmann
