+++ docs/sphinx/source/reference/pv_modeling/index.rst
@@ -25,4 +25,3 @@ wrap the functions listed below. See its documentation for details.
    system_models
    parameters
    other
-   windspeed
