--- docs/sphinx/source/reference/pv_modeling/windspeed.rst
@@ -1,9 +1,0 @@
-.. currentmodule:: pvlib
-
-Wind Speed
-----------
-
-.. autosummary::
-   :toctree: ../generated/
-
-   windspeed.wind_speed_at_height_hellmann
